==> modules/8_2_plot_update_signals_ratio_module_chr.R <==
###############################################################
########### plot updated unique significant SNPs ###########
###############################################################
# load packages -----
rm(list = ls())
library(data.table)
library(tidyverse)
source("~/Trans/plot/theme_my_pub.R")


# I/O & paras -----
ratio_thre <- 50
file_signal <- paste0('postanalysis/signal_rm_infl_ratio_', ratio_thre, '_module_chr.txt')

## output -----
file_plot_module <- paste0('plot/signal_module_rm_infl_ratio_', ratio_thre, '_module_chr.pdf')
file_plot_chr <- paste0('plot/signal_chr_rm_infl_ratio_', ratio_thre, '_module_chr.pdf')
file_plot_module_chr <- paste0('plot/signal_module_chr_rm_infl_ratio_', ratio_thre, '_module_chr.pdf')


# read files -----
signal <- fread(file_signal)


# how many signals each module has?
signal %>% count(module) %>% arrange(desc(n))
signal %>% count(module, SNPChr) %>% arrange(desc(n))


# how many modules each signal correspond to?
signal %>%
  group_by(SNP, meta, SNPChr, SNPPos) %>%
  summarise("n_module" = n(), "minp" = min(p) ) %>%
  arrange(minp, desc(n_module), SNPChr)


# 1. plot how many signals each module has? -----
signal$module <- factor(signal$module)

fig <- ggplot(signal, aes(x = module)) +
  geom_bar(stat = "count") +
  labs(x = "Module", y = "Number of signals\n(SNPs)")
fig +
  theme_my_pub() +
  theme(axis.text.x = element_text(size = 6, angle = 90),
        axis.ticks = element_blank(),
        axis.line.y = element_blank(),
        panel.grid.major.y = element_line(linetype = "dashed"))

ggsave(file_plot_module, width = 5, height = 4)


# 2. plt how many signals are on each chr? -----
signal$SNPChr <- factor(signal$SNPChr)

fig <- ggplot(signal %>% distinct(SNP, .keep_all = TRUE),
              aes(x = SNPChr)) +
  geom_bar(stat = "count") +
  labs(x = "Chromosome", y = "Number of signals\n(SNPs)")
fig +
  theme_my_pub() +
  theme(axis.text.x = element_text(angle = 60),
        axis.ticks = element_blank(),
        axis.line.y = element_blank(),
        panel.grid.major.y = element_line(linetype = "dashed"))

ggsave(file_plot_chr, width = 6, height = 4)


# 3. plt how many signals for each chr and module? -----
# set module & chr order on the figure
plt_df <- signal %>%
  group_by(module, SNPChr) %>%
  summarise(n_M_chr = n()) %>%
  ungroup() %>%
  arrange(desc(n_M_chr))
plt_df$module = factor(plt_df$module, levels = plt_df$module, labels = plt_df$module)
plt_df$SNPChr = factor(plt_df$SNPChr, levels = (plt_df$SNPChr) %>% unique() %>% rev())


fig <- ggplot(plt_df, aes(x = module, y = SNPChr)) +
  geom_tile(aes(fill = n_M_chr)) +
  scale_fill_gradientn(colors = RColorBrewer::brewer.pal(8, "Blues")[3:8]) +
  labs(x = "Module", y = "Chromosome", fill = "SNPs")
fig +
  theme_my_pub(legend.text.size = 10, legend.position = "right") +
  theme(axis.text.x = element_text(angle = 90, size = 6),
        axis.ticks.x = element_blank())

ggsave(file_plot_module_chr, width = 5, height = 5)
